requirements/requirements.txt
django==1.7
psycopg2==2.5.4
django-extensions
django-bootstrap3
django-debug-toolbar
django-crispy-forms
django-bower
django-redis
djangorestframework
markdown
celery
ipdb
nltk
